+++ send_email_messages/frontend/util.js
@@ -5,6 +5,7 @@ export function Accordion({title, children}) {
   const [isOpen, setIsOpen] = useState(false);
   return (<Box>
     <TextButton 
+      marginY={1}
       onClick={() => setIsOpen(!isOpen)} 
       icon={isOpen ? "chevronDown" : "chevronRight"}>
       {title}
