+++ send_email_messages/package.json
@@ -2,7 +2,8 @@
     "dependencies": {
         "@airtable/blocks": "0.0.48",
         "react": "^16.13.0",
-        "react-dom": "^16.13.0"
+        "react-dom": "^16.13.0",
+        "react-step-wizard": "^5.3.2"
     },
     "devDependencies": {
         "eslint": "^6.8.0",
